refactor: 폴더 병합 및 불필요한 코드 삭제
