+++ src/main/java/com/management/chatbot/repository/MemberRepository.java
@@ -9,4 +9,5 @@ import java.util.Optional;
 @Repository
 public interface MemberRepository extends JpaRepository<Member, Long> {
     Member findByKakaoId(String kakaoId);
+    Member save(Member member);
 }
