+++ src/main/java/com/management/chatbot/service/dto/KakaoDto/KakaoRequestDto.java
@@ -3,44 +3,17 @@ package com.management.chatbot.service.dto.KakaoDto;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.util.List;
 import java.util.Map;
 
 @Getter
 @NoArgsConstructor
 public class KakaoRequestDto {
-    private Intent intent;
-    private UserRequest userRequest;
     private Bot bot;
+    private Intent intent;
     private Action action;
-
-    @Getter
-    public static class Intent {
-        private String id;
-        private String name;
-    }
-
-    @Getter
-    public static class UserRequest {
-        private String timezone;
-        private Map<String, String> params;
-        private Block block;
-        private String utterance;
-        private String lang;
-        private User user;
-    }
-
-    @Getter
-    public static class Block {
-        private String id;
-        private String name;
-    }
-
-    @Getter
-    public static class User {
-        private String id;
-        private String type;
-        private Map<String, Object> properties;
-    }
+    private UserRequest userRequest;
+    private List<Object> contexts;
 
     @Getter
     public static class Bot {
@@ -49,18 +22,78 @@ public class KakaoRequestDto {
     }
 
     @Getter
-    public static class Action {
+    public static class Intent {
         private String id;
         private String name;
-        private Map<String, String> clientExtra;
-        private Map<String, String> params;
-        private Map<String, DetailParam> detailParams;
+        private Extra extra;
     }
 
     @Getter
-    public static class DetailParam {
+    public static class Extra {
+        private Reason reason;
+        public static class Reason {
+            private int code;
+            private String message;
+        }
+    }
+
+    @Getter
+    public static class Action {
+        private String id;
+        private String name;
+        private Map<String, String> params;
+        private Map<String, DetailParams> detailParams;
+        private Map<String, String> clientExtra;
+    }
+
+    @Getter
+    public static class DetailParams {
         private String groupName;
         private String origin;
         private String value;
+    }
+
+    @Getter
+    public static class UserRequest {
+        private Block block;
+        private User user;
+        private String utterance;
+        private Params params;
+        private String lang;
+        private String timezone;
+
+        @Getter
+        public static class Block {
+            private String id;
+            private String name;
+        }
+
+        @Getter
+        public static class User {
+            private String id;
+            private String type;
+            private Properties properties;
+
+            @Getter
+            public static class Properties {
+                private String botUserKey;
+                private boolean isFriend;
+                private String plusfriendUserKey;
+                private String bot_user_key;
+                private String plusfriend_user_key;
+            }
+        }
+
+        @Getter
+        public static class Params {
+            private String surface;
+            private Media media;
+
+            @Getter
+            public static class Media {
+                private String type;
+                private String url;
+            }
+        }
     }
 }
