+++ src/main/java/com/management/chatbot/service/dto/KakaoDto/KakaoResponseDto.java
@@ -2,10 +2,12 @@ package com.management.chatbot.service.dto.KakaoDto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @Getter
 @NoArgsConstructor
@@ -39,5 +41,29 @@ public class KakaoResponseDto {
         resultJson.put("data", data);
 
         return resultJson;
+    }
+
+    @Getter @Setter
+    @NoArgsConstructor
+    public static class KakaoBasicCardResponseDto {
+        public HashMap<String, Object> makeResponseBody(List<HashMap<String, Object>> values) {
+
+            HashMap<String, Object> resultJson = new HashMap<>();
+
+            List<HashMap<String, Object>> outputs = new ArrayList<>();
+            HashMap<String, Object> template = new HashMap<>();
+            HashMap<String, Object> basicCard = new HashMap<>();
+
+            for (HashMap<String, Object> value : values) {
+                outputs.add(value);
+            }
+
+            template.put("outputs", outputs);
+
+            resultJson.put("version", "2.0");
+            resultJson.put("template", template);
+
+            return resultJson;
+        }
     }
 }
