+++ src/main/java/com/management/web/repository/GiftcardOrderRepository.java
@@ -3,6 +3,8 @@ package com.management.web.repository;
 
 import com.management.web.domain.GiftcardOrder;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public interface GiftcardOrderRepository extends JpaRepository<GiftcardOrder, Long> {
 }
