+++ src/main/java/com/management/web/repository/GiftcardProductRepository.java
@@ -3,9 +3,11 @@ package com.management.web.repository;
 import com.management.web.domain.GiftcardProduct;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
+import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
+@Repository
 public interface GiftcardProductRepository extends JpaRepository<GiftcardProduct, Long> {
 
     List<GiftcardProduct> findBySaleYN(Boolean saleYN);
