+++ src/main/java/com/management/web/service/MainService.java
@@ -1,7 +1,7 @@
-package com.management.chatbot.web.service;
+package com.management.web.service;
 
-import com.management.chatbot.web.dto.MyMainInfoDto;
-import com.management.chatbot.web.repository.MemberWebRepository;
+import com.management.web.service.dto.MyMainInfoDto;
+import com.management.web.repository.MemberWebRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
